Dev jj 3.2

+++ myWebApp/Pages/AddWorkspace.cshtml.cs
@@ -36,7 +36,7 @@ namespace myWebApp.Pages
             CreateWorkspace(workspace.LocationName, workspace.RoomName, workspace.SquareMeters, availworkspaces);
         }
 
-        public int AvailableWorkspaces(int SquareMeters, int x, int y)
+        public int AvailableWorkspaces(int SquareMeters, double x, double y)
         {
             double X = 1.5 + x;
             double Y = 1.5 + y;
@@ -67,6 +67,7 @@ namespace myWebApp.Pages
             cmd.ExecuteNonQuery();
             con.Close();
         }
+
         public List<Workspace> ShowWorkspaces()
         {
             List<Workspace> Workspaces = new List<Workspace>();
@@ -87,5 +88,6 @@ namespace myWebApp.Pages
             }
             return Workspaces;
         }
+
     }
 }
